requirements.txt
Flask==1.1.2
google-cloud-storage==1.36.2
spleeter==2.2.1
redis==3.5.3
gunicorn==20.0.4